requirements.txt
fastapi
uvicorn
postgres
routers
pydantic==1.10.11
pydantic_settings
sqlalchemy
psycopg2
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
